requirements.txt
flask==2.3.2
flask-cors==4.0.0
supabase==2.16.0
python-dotenv==1.1.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.2.2
gunicorn==23.0.0
setuptools>=65.5.0